+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.1.0
+
+* `Knp\Menu\Twig\Helper` is deprecated in favor of `Knp\Menu\Templating\Helper`
+* `Knp\Menu\Twig\MenuExtension` is deprecated in favor of `Knp\Menu\Integration\Twig\MenuExtension`
+* `Knp\Menu\Renderer\TwigRenderer` is deprecated in favor of `Knp\Menu\Integration\Twig\TwigRenderer`
+
 ## 2.0.1 (2014-08-01)
 
 * Fixed voter conventions on RouteVoter